fintools-agent-client/requirements.txt
python-dotenv==1.2.2
a2a-sdk[http-server]==0.3.24
